Refactor metacat ini order to ensure controllers are inited after the catalog
