+++ metacat-main/src/main/java/com/netflix/metacat/main/api/v1/MetadataController.java
@@ -32,9 +32,11 @@ import com.netflix.metacat.main.api.RequestWrapper;
 import com.netflix.metacat.main.services.GetTableServiceParameters;
 import com.netflix.metacat.main.services.MetacatServiceHelper;
 import com.netflix.metacat.main.services.MetadataService;
+import com.netflix.metacat.main.services.init.MetacatCoreInitService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -66,32 +68,14 @@ import java.util.Set;
     produces = MediaType.APPLICATION_JSON_VALUE,
     consumes = MediaType.APPLICATION_JSON_VALUE
 )
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class MetadataController {
     private final UserMetadataService userMetadataService;
     private final MetacatServiceHelper helper;
     private final MetadataService metadataService;
     private final RequestWrapper requestWrapper;
-
-    /**
-     * Constructor.
-     *
-     * @param userMetadataService user metadata service
-     * @param helper              helper
-     * @param metadataService     metadata service
-     * @param requestWrapper      request wrapper object
-     */
-    @Autowired
-    public MetadataController(
-        final UserMetadataService userMetadataService,
-        final MetacatServiceHelper helper,
-        final MetadataService metadataService,
-        final RequestWrapper requestWrapper
-    ) {
-        this.userMetadataService = userMetadataService;
-        this.helper = helper;
-        this.metadataService = metadataService;
-        this.requestWrapper = requestWrapper;
-    }
+    // this is needed to ensure everything is initialized before the controllers are
+    private final MetacatCoreInitService metacatCoreInitService;
 
     /**
      * Returns the data metadata.
