+++ metacat-main/src/main/java/com/netflix/metacat/main/api/v1/SearchController.java
@@ -18,9 +18,11 @@
 package com.netflix.metacat.main.api.v1;
 
 import com.netflix.metacat.common.dto.TableDto;
-import com.netflix.metacat.main.services.search.ElasticSearchUtil;
 import com.netflix.metacat.main.api.RequestWrapper;
+import com.netflix.metacat.main.services.init.MetacatCoreInitService;
+import com.netflix.metacat.main.services.search.ElasticSearchUtil;
 import io.swagger.annotations.ApiParam;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.HttpStatus;
@@ -42,24 +44,12 @@ import java.util.List;
     path = "/mds/v1/search",
     produces = MediaType.APPLICATION_JSON_VALUE
 )
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class SearchController {
     private final ElasticSearchUtil elasticSearchUtil;
     private final RequestWrapper requestWrapper;
-
-    /**
-     * Constructor.
-     *
-     * @param elasticSearchUtil search util
-     * @param requestWrapper    request wrapper object
-     */
-    @Autowired
-    public SearchController(
-        final ElasticSearchUtil elasticSearchUtil,
-        final RequestWrapper requestWrapper
-    ) {
-        this.elasticSearchUtil = elasticSearchUtil;
-        this.requestWrapper = requestWrapper;
-    }
+    // this is needed to ensure everything is initialized before the controllers are
+    private final MetacatCoreInitService metacatCoreInitService;
 
     /**
      * Searches the list of tables for the given search string.
